Pulled logic out of ProfilePageController
