+++ src/Foundation/Features/MyAccount/ProfilePage/ProfilePageController.cs
@@ -1,18 +1,8 @@
-﻿using EPiServer;
-using EPiServer.Cms.UI.AspNetIdentity;
-using EPiServer.Commerce.Order;
-using EPiServer.Core;
-using EPiServer.Security;
-using EPiServer.Web.Routing;
+﻿using EPiServer.Cms.UI.AspNetIdentity;
 using Foundation.Cms.Identity;
 using Foundation.Commerce.Customer.Services;
 using Foundation.Commerce.Customer.ViewModels;
-using Foundation.Commerce.Models.Pages;
-using Foundation.Commerce.Order.Services;
-using Foundation.Commerce.Order.ViewModels;
-using Mediachase.Commerce.Security;
-using System.Collections.Generic;
-using System.Linq;
+using System;
 using System.Web.Mvc;
 
 namespace Foundation.Features.MyAccount.ProfilePage
@@ -20,36 +10,19 @@ namespace Foundation.Features.MyAccount.ProfilePage
     [Authorize]
     public class ProfilePageController : IdentityControllerBase<Social.Models.Pages.ProfilePage>
     {
-        private readonly IAddressBookService _addressBookService;
-        private readonly IOrderRepository _orderRepository;
-        private readonly IContentLoader _contentLoader;
-        private readonly ICartService _cartService;
+        private readonly ProfilePageControllerService _controllerService;
 
-        public ProfilePageController(IAddressBookService addressBookService,
-            IOrderRepository orderRepository,
+        public ProfilePageController(ProfilePageControllerService controllerService,
             ApplicationSignInManager<SiteUser> signinManager,
             ApplicationUserManager<SiteUser> userManager,
-            ICartService cartService,
-            IContentLoader contentLoader,
             ICustomerService customerService) : base(signinManager, userManager, customerService)
         {
-            _addressBookService = addressBookService;
-            _orderRepository = orderRepository;
-            _contentLoader = contentLoader;
-            _cartService = cartService;
+            _controllerService = controllerService ?? throw new ArgumentNullException(nameof(controllerService));
         }
 
         public ActionResult Index(Social.Models.Pages.ProfilePage currentPage)
         {
-            var viewModel = new SocialCommerceProfilePageViewModel(currentPage)
-            {
-                Orders = GetOrderHistoryViewModels(),
-                Addresses = GetAddressViewModels(),
-                SiteUser = CustomerService.GetSiteUser(User.Identity.Name),
-                CustomerContact = new Commerce.Customer.FoundationContact(CustomerService.GetCurrentContact().Contact)
-            };
-
-            viewModel.OrderDetailsPageUrl = UrlResolver.Current.GetUrl(_contentLoader.Get<CommerceHomePage>(ContentReference.StartPage).OrderDetailsPage);
+            var viewModel = _controllerService.GetViewModel(currentPage, User.Identity.Name, CustomerService);
 
             return View(viewModel);
         }
@@ -72,55 +45,6 @@ namespace Foundation.Features.MyAccount.ProfilePage
             contact.SaveChanges();
 
             return Json(new { FirstName = contact.FirstName, LastName = contact.LastName });
-        }
-
-        private IList<AddressModel> GetAddressViewModels() => _addressBookService.List();
-
-        private List<OrderViewModel> GetOrderHistoryViewModels()
-        {
-            var purchaseOrders = _orderRepository.Load<IPurchaseOrder>(PrincipalInfo.CurrentPrincipal.GetContactId(), _cartService.DefaultCartName)
-                .OrderByDescending(x => x.Created).ToList();
-
-            if (purchaseOrders.Count > 3)
-            {
-                purchaseOrders = purchaseOrders.Take(3).ToList();
-            }
-
-
-            var viewModel = new List<OrderViewModel>();
-
-            foreach (var purchaseOrder in purchaseOrders)
-            {
-                // Assume there is only one form per purchase.
-                var form = purchaseOrder.GetFirstForm();
-                var billingAddress = new AddressModel();
-                var payment = form.Payments.FirstOrDefault();
-                if (payment != null)
-                {
-                    billingAddress = _addressBookService.ConvertToModel(payment.BillingAddress);
-                }
-                var orderViewModel = new OrderViewModel
-                {
-                    PurchaseOrder = purchaseOrder,
-                    Items = form.GetAllLineItems().Select(lineItem => new OrderHistoryItemViewModel
-                    {
-                        LineItem = lineItem,
-                    }).GroupBy(x => x.LineItem.Code).Select(group => group.First()),
-                    BillingAddress = billingAddress,
-                    ShippingAddresses = new List<AddressModel>(),
-                    OrderGroupId = purchaseOrder.OrderLink.OrderGroupId
-                };
-
-                foreach (var orderAddress in purchaseOrder.Forms.SelectMany(x => x.Shipments).Select(s => s.ShippingAddress))
-                {
-                    var shippingAddress = _addressBookService.ConvertToModel(orderAddress);
-                    orderViewModel.ShippingAddresses.Add(shippingAddress);
-                }
-
-                viewModel.Add(orderViewModel);
-            }
-
-            return viewModel;
         }
     }
 }
