+++ src/Foundation/Foundation.csproj
@@ -867,6 +867,7 @@
     <Compile Include="Features\Blog\TagCloudBlock\TagCloudBlockController.cs" />
     <Compile Include="Features\Blog\TagCloudBlock\TagCloudBlockControllerService.cs" />
     <Compile Include="Features\MyAccount\OrderHistory\OrderHistoryControllerService.cs" />
+    <Compile Include="Features\MyAccount\ProfilePage\ProfilePageControllerService.cs" />
     <Compile Include="Infrastructure\PostClickInterceptor.cs" />
     <Compile Include="Features\CatalogContent\CatalogContentControllerBase.cs" />
     <Compile Include="Features\CatalogContent\Variation\VariationController.cs" />
